+++ src/js/chrome/chrome_shim.js
@@ -203,131 +203,6 @@ var chromeShim = {
     });
   },
 
-  shimGetUserMedia: function() {
-    var constraintsToChrome_ = function(c) {
-      if (typeof c !== 'object' || c.mandatory || c.optional) {
-        return c;
-      }
-      var cc = {};
-      Object.keys(c).forEach(function(key) {
-        if (key === 'require' || key === 'advanced' || key === 'mediaSource') {
-          return;
-        }
-        var r = (typeof c[key] === 'object') ? c[key] : {ideal: c[key]};
-        if (r.exact !== undefined && typeof r.exact === 'number') {
-          r.min = r.max = r.exact;
-        }
-        var oldname_ = function(prefix, name) {
-          if (prefix) {
-            return prefix + name.charAt(0).toUpperCase() + name.slice(1);
-          }
-          return (name === 'deviceId') ? 'sourceId' : name;
-        };
-        if (r.ideal !== undefined) {
-          cc.optional = cc.optional || [];
-          var oc = {};
-          if (typeof r.ideal === 'number') {
-            oc[oldname_('min', key)] = r.ideal;
-            cc.optional.push(oc);
-            oc = {};
-            oc[oldname_('max', key)] = r.ideal;
-            cc.optional.push(oc);
-          } else {
-            oc[oldname_('', key)] = r.ideal;
-            cc.optional.push(oc);
-          }
-        }
-        if (r.exact !== undefined && typeof r.exact !== 'number') {
-          cc.mandatory = cc.mandatory || {};
-          cc.mandatory[oldname_('', key)] = r.exact;
-        } else {
-          ['min', 'max'].forEach(function(mix) {
-            if (r[mix] !== undefined) {
-              cc.mandatory = cc.mandatory || {};
-              cc.mandatory[oldname_(mix, key)] = r[mix];
-            }
-          });
-        }
-      });
-      if (c.advanced) {
-        cc.optional = (cc.optional || []).concat(c.advanced);
-      }
-      return cc;
-    };
-
-    var getUserMedia_ = function(constraints, onSuccess, onError) {
-      if (constraints.audio) {
-        constraints.audio = constraintsToChrome_(constraints.audio);
-      }
-      if (constraints.video) {
-        constraints.video = constraintsToChrome_(constraints.video);
-      }
-      logging('chrome: ' + JSON.stringify(constraints));
-      return navigator.webkitGetUserMedia(constraints, onSuccess, onError);
-    };
-    navigator.getUserMedia = getUserMedia_;
-
-    // Returns the result of getUserMedia as a Promise.
-    var getUserMediaPromise_ = function(constraints) {
-      return new Promise(function(resolve, reject) {
-        navigator.getUserMedia(constraints, resolve, reject);
-      });
-    }
-
-    if (!navigator.mediaDevices) {
-      navigator.mediaDevices = {getUserMedia: getUserMediaPromise_,
-                               enumerateDevices: function() {
-        return new Promise(function(resolve) {
-          var kinds = {audio: 'audioinput', video: 'videoinput'};
-          return MediaStreamTrack.getSources(function(devices) {
-            resolve(devices.map(function(device) {
-              return {label: device.label,
-                      kind: kinds[device.kind],
-                      deviceId: device.id,
-                      groupId: ''};
-            }));
-          });
-        });
-      }};
-    }
-
-    // A shim for getUserMedia method on the mediaDevices object.
-    // TODO(KaptenJansson) remove once implemented in Chrome stable.
-    if (!navigator.mediaDevices.getUserMedia) {
-      navigator.mediaDevices.getUserMedia = function(constraints) {
-        return getUserMediaPromise_(constraints);
-      };
-    } else {
-      // Even though Chrome 45 has navigator.mediaDevices and a getUserMedia
-      // function which returns a Promise, it does not accept spec-style
-      // constraints.
-      var origGetUserMedia = navigator.mediaDevices.getUserMedia.
-          bind(navigator.mediaDevices);
-      navigator.mediaDevices.getUserMedia = function(c) {
-        if (c) {
-          logging('spec:   ' + JSON.stringify(c)); // whitespace for alignment
-          c.audio = constraintsToChrome_(c.audio);
-          c.video = constraintsToChrome_(c.video);
-          logging('chrome: ' + JSON.stringify(c));
-        }
-        return origGetUserMedia(c);
-      }.bind(this);
-    }
-
-    // Dummy devicechange event methods.
-    // TODO(KaptenJansson) remove once implemented in Chrome stable.
-    if (typeof navigator.mediaDevices.addEventListener === 'undefined') {
-      navigator.mediaDevices.addEventListener = function() {
-        logging('Dummy mediaDevices.addEventListener called.');
-      };
-    }
-    if (typeof navigator.mediaDevices.removeEventListener === 'undefined') {
-      navigator.mediaDevices.removeEventListener = function() {
-        logging('Dummy mediaDevices.removeEventListener called.');
-      };
-    }
-  },
-
   // Attach a media stream to an element.
   attachMediaStream: function(element, stream) {
     logging('DEPRECATED, attachMediaStream will soon be removed.');
@@ -350,12 +225,13 @@ var chromeShim = {
   }
 }
 
+
 // Expose public methods.
 module.exports = {
   shimOnTrack: chromeShim.shimOnTrack,
   shimSourceObject: chromeShim.shimSourceObject,
   shimPeerConnection: chromeShim.shimPeerConnection,
-  shimGetUserMedia: chromeShim.shimGetUserMedia,
+  shimGetUserMedia: require('./getusermedia'),
   attachMediaStream: chromeShim.attachMediaStream,
   reattachMediaStream: chromeShim.reattachMediaStream
 };
